refactor docolon to prepare for execute word
Before the change, docolon assumed that the previous instruction was
hardcoded four bytes before the current instruction. With adding the
possibility for interpreted words, this is not guaranteed anymore: A
word can be fetched from the dictionary instead.

The responsibility for managing the correct address of the next forth
instruction now is on the caller of docolon. docolon expects an address
on the stack. `call docolon` puts this exact address.